backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.5
postgrest-py==0.13.0
realtime-py==1.0.0
storage-py==0.7.0
jwt==1.3.0
python-multipart==0.0.6
httpx==0.25.2
slowapi==0.1.9
